move Setup class to own file in src/app/setup folder
