+++ src/app/model.ts
@@ -304,15 +304,6 @@ export interface OAuthApp extends BaseEntity {
 };
 export type OAuthApps = Array < OAuthApp > ;
 
-export interface Setup {
-  gitHubOAuthConfiguration: GitHubOAuthConfiguration;
-}
-
-export interface GitHubOAuthConfiguration {
-  clientId: string;
-  clientSecret: string;
-}
-
 export interface ListResultString extends BaseEntity {
     items: Array < string >
     ;
