+++ src/app/setup/setup.ts
@@ -1,0 +1,8 @@
+export interface Setup {
+  gitHubOAuthConfiguration: GitHubOAuthConfiguration;
+}
+
+export interface GitHubOAuthConfiguration {
+  clientId: string;
+  clientSecret: string;
+}
